abstract out deleteFrame
